fix(#2413): move progress bar in breakdown.tsx
