+++ libs/accounts/src/lib/breakdown-table.tsx
@@ -5,19 +5,47 @@ import {
   PriceCell,
   t,
 } from '@vegaprotocol/react-helpers';
-import type { VegaValueFormatterParams } from '@vegaprotocol/ui-toolkit';
 import {
   AgGridDynamic as AgGrid,
+  Intent,
   progressBarCellRendererSelector,
 } from '@vegaprotocol/ui-toolkit';
 import { AgGridColumn } from 'ag-grid-react';
 import type { AgGridReact, AgGridReactProps } from 'ag-grid-react';
 import type { AccountFields } from './accounts-data-provider';
 import { AccountTypeMapping } from '@vegaprotocol/types';
-import {
-  progressBarHeaderComponentParams,
-  progressBarValueFormatter,
-} from './accounts-table';
+import type {
+  ValueProps,
+  VegaValueFormatterParams,
+} from '@vegaprotocol/ui-toolkit';
+import type { ValueFormatterParams } from 'ag-grid-community';
+
+export const progressBarValueFormatter = ({
+  data,
+  node,
+}: ValueFormatterParams): ValueProps['valueFormatted'] | undefined => {
+  if (!data || node?.rowPinned) {
+    return undefined;
+  }
+  const min = BigInt(data.used);
+  const mid = BigInt(data.available);
+  const max = BigInt(data.deposited);
+  const range = max > min ? max : min;
+  return {
+    low: addDecimalsFormatNumber(min.toString(), data.asset.decimals, 4),
+    high: addDecimalsFormatNumber(mid.toString(), data.asset.decimals, 4),
+    value: range ? Number((min * BigInt(100)) / range) : 0,
+    intent: Intent.Warning,
+  };
+};
+
+export const progressBarHeaderComponentParams = {
+  template:
+    '<div class="ag-cell-label-container" role="presentation">' +
+    `  <span>${t('Available')}</span>` +
+    '  <span ref="eText" class="ag-header-cell-text"></span>' +
+    '</div>',
+};
 
 interface BreakdownTableProps extends AgGridReactProps {
   data: AccountFields[] | null;
